♻️ Refactor Modal to Header, Contents and Footer
